Extract remote-status payload, sidebar remote help, and CLI version lookup into new files
Keeps Workspace.swift, ContentView.swift, and cmux.swift below their merge-base
line counts (hard-cap ratchet in swift_file_length_budget.py --base-ref), and
switches the sidebar help strings to interpolated String(localized:).
