+++ CLI/cmux.swift
@@ -35034,89 +35034,6 @@ export default CMUXSessionRestore;
         return String(normalized.prefix(12))
     }
 
-    // Foundation can walk past "/" into "/.." when repeatedly deleting path
-    // components, so stop once the canonical root is reached.
-    func parentSearchURL(for url: URL) -> URL? {
-        let standardized = url.standardizedFileURL
-        let path = standardized.path
-        guard !path.isEmpty, path != "/" else {
-            return nil
-        }
-
-        let parent = standardized.deletingLastPathComponent().standardizedFileURL
-        guard parent.path != path else {
-            return nil
-        }
-        return parent
-    }
-
-    func candidateInfoPlistURLs() -> [URL] {
-        guard let executableURL = resolvedExecutableURL() else {
-            return []
-        }
-
-        let fileManager = FileManager.default
-
-        var candidates: [URL] = []
-        var seen: Set<String> = []
-        func appendIfExisting(_ url: URL) {
-            let path = url.path
-            guard !path.isEmpty else { return }
-            guard seen.insert(path).inserted else { return }
-            guard fileManager.fileExists(atPath: path) else { return }
-            candidates.append(url)
-        }
-
-        var current = executableURL.deletingLastPathComponent().standardizedFileURL
-        while true {
-            if current.pathExtension == "app" {
-                appendIfExisting(current.appendingPathComponent("Contents/Info.plist"))
-            }
-            if current.lastPathComponent == "Contents" {
-                appendIfExisting(current.appendingPathComponent("Info.plist"))
-            }
-
-            let projectMarker = current.appendingPathComponent("cmux.xcodeproj/project.pbxproj")
-            let repoInfo = current.appendingPathComponent("Resources/Info.plist")
-            if fileManager.fileExists(atPath: projectMarker.path),
-               fileManager.fileExists(atPath: repoInfo.path) {
-                appendIfExisting(repoInfo)
-                break
-            }
-
-            guard let parent = parentSearchURL(for: current) else {
-                break
-            }
-            current = parent
-        }
-
-        // If we already found an ancestor bundle or repo Info.plist, avoid scanning
-        // sibling app bundles. Large Resources directories can otherwise balloon RSS.
-        guard candidates.isEmpty else {
-            return candidates
-        }
-
-        let searchRoots = [
-            executableURL.deletingLastPathComponent().standardizedFileURL,
-            executableURL.deletingLastPathComponent().deletingLastPathComponent().standardizedFileURL
-        ]
-        for root in searchRoots {
-            guard let entries = fileManager.enumerator(
-                at: root,
-                includingPropertiesForKeys: nil,
-                options: [.skipsHiddenFiles, .skipsSubdirectoryDescendants],
-                errorHandler: { _, _ in true }
-            ) else {
-                continue
-            }
-            for case let entry as URL in entries where entry.pathExtension == "app" {
-                appendIfExisting(entry.appendingPathComponent("Contents/Info.plist"))
-            }
-        }
-
-        return candidates
-    }
-
     private func currentExecutablePath() -> String? {
         if let path = CLIExecutableLocator.currentExecutableURL()?.path
             .trimmingCharacters(in: .whitespacesAndNewlines),
